add mini tetris to menu
