+++ index.html
@@ -9,6 +9,6 @@
 </head>
 <body>
 	<canvas id="myCanvas"></canvas>
-	<script type="text/javascript" src="./js/main.js"></script>
+	<script type="module" src="./js/main.js"></script>
 </body>
 </html>
